+++ generators/app/templates/infrastructure/src/components/layout/header/WebHeader.js
@@ -1,29 +1,34 @@
-import React, { useRef } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import { useTranslation } from 'react-i18next'
 import { useHeader } from 'providers/AreasProvider'
-import { HeaderRef, HeaderResponsive, StyledMobileTitle } from './HeaderStyle'
+import { useTranslation } from 'react-i18next'
+import { MinimizeSidebar, StyledTitle } from './HeaderStyle'
 import { IconButton } from '@totalsoft/rocket-ui'
-import Menu from '@mui/icons-material/Menu'
+import { MoreVert, ViewList } from '@mui/icons-material'
 
-const WebHeader = ({ handleDrawerToggle, pathName }) => {
+const WebHeader = ({ drawerOpen, handleDrawerToggle, pathName }) => {
   const { t } = useTranslation()
   const [header] = useHeader()
-  const headerRef = useRef()
   return (
     <>
-      {header || <StyledMobileTitle variant='subtitle1'>{pathName && t('NavBar.' + pathName)}</StyledMobileTitle>}
-      <HeaderRef ref={headerRef}></HeaderRef>
-      <HeaderResponsive>
-        <IconButton color='secondary' aria-label='open drawer' onClick={handleDrawerToggle}>
-          <Menu fontSize='small' />
-        </IconButton>
-      </HeaderResponsive>
+      <MinimizeSidebar>
+        {drawerOpen ? (
+          <IconButton color='white' onClick={handleDrawerToggle} size='medium'>
+            <MoreVert fontSize='small' />
+          </IconButton>
+        ) : (
+          <IconButton color='white' onClick={handleDrawerToggle} size='medium'>
+            <ViewList fontSize='small' />
+          </IconButton>
+        )}
+      </MinimizeSidebar>
+      {header || <StyledTitle variant='subtitle1'>{pathName && t('NavBar.' + pathName)}</StyledTitle>}
     </>
   )
 }
 
 WebHeader.propTypes = {
+  drawerOpen: PropTypes.bool.isRequired,
   handleDrawerToggle: PropTypes.func,
   pathName: PropTypes.string
 }
